Clean and move bilinear_tensor_product
